Add Biome for linting and formatting
