+++ cli.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env bun
-import {Database} from 'bun:sqlite'
+import { Database } from 'bun:sqlite'
 
 const API = 'https://api.slaytheweb.cards'
 const db = new Database('runs.db')
@@ -22,18 +22,31 @@ function table(rows: Record<string, unknown>[]) {
 
 const commands: Record<string, (arg?: string) => void | Promise<void>> = {
 	async pull() {
-		const lastId = db.query('select max(id) as id from runs').get() as {id: number | null}
+		const lastId = db.query('select max(id) as id from runs').get() as {
+			id: number | null
+		}
 		const localMax = lastId?.id || 0
 		console.log(`Local: ${localMax} runs`)
 
 		const res = await fetch(`${API}/api/runs`)
-		const {total, runs} = (await res.json()) as {total: number; runs: {id: number; player: string; createdAt: number; endedAt: number; floor: number; won: number}[]}
+		const { total, runs } = (await res.json()) as {
+			total: number
+			runs: {
+				id: number
+				player: string
+				createdAt: number
+				endedAt: number
+				floor: number
+				won: number
+			}[]
+		}
 		const fresh = runs.filter((r) => r.id > localMax)
 
 		if (!fresh.length) return console.log('Up to date.')
 
 		const ins = db.prepare('insert or ignore into runs values (?,?,?,?,?,?)')
-		for (const r of fresh) ins.run(r.id, r.player, r.createdAt, r.endedAt, r.floor, r.won)
+		for (const r of fresh)
+			ins.run(r.id, r.player, r.createdAt, r.endedAt, r.floor, r.won)
 
 		console.log(`+${fresh.length} runs (${total} total)`)
 	},
@@ -44,25 +57,31 @@ const commands: Record<string, (arg?: string) => void | Promise<void>> = {
 	},
 
 	stats() {
-		const rows = db.query(`select
+		const rows = db
+			.query(`select
 			count(*) as total,
 			sum(won) as wins,
 			round(100.0 * sum(won) / count(*), 1) as 'win%',
 			round(avg(floor), 1) as avg_floor,
 			round(avg(ended_at - created_at) / 1000) as avg_sec
-		from runs`).all() as Record<string, unknown>[]
+		from runs`)
+			.all() as Record<string, unknown>[]
 		table(rows)
 	},
 
 	top() {
-		const rows = db.query(`select player, count(*) as runs, sum(won) as wins
-			from runs group by player order by runs desc limit 10`).all() as Record<string, unknown>[]
+		const rows = db
+			.query(`select player, count(*) as runs, sum(won) as wins
+			from runs group by player order by runs desc limit 10`)
+			.all() as Record<string, unknown>[]
 		table(rows)
 	},
 
 	daily() {
-		const rows = db.query(`select date(created_at/1000, 'unixepoch') as day, count(*) as n
-			from runs group by day order by day desc limit 14`).all() as Record<string, unknown>[]
+		const rows = db
+			.query(`select date(created_at/1000, 'unixepoch') as day, count(*) as n
+			from runs group by day order by day desc limit 14`)
+			.all() as Record<string, unknown>[]
 		table(rows)
 	},
 
@@ -72,34 +91,55 @@ const commands: Record<string, (arg?: string) => void | Promise<void>> = {
 
 		// Get a real run ID first
 		const runsRes = await fetch(`${base}/api/runs`)
-		const runsData = (await runsRes.json()) as {runs: {id: number}[]}
+		const runsData = (await runsRes.json()) as { runs: { id: number }[] }
 		const testId = runsData.runs[0]?.id || 1
 
+		function isObject(d: unknown): d is Record<string, unknown> {
+			return typeof d === 'object' && d !== null
+		}
+
 		const tests = [
-			{name: 'GET /api/runs', path: '/api/runs', check: (d: any) => typeof d.total === 'number' && Array.isArray(d.runs)},
-			{name: `GET /api/runs/${testId}`, path: `/api/runs/${testId}`, check: (d: any) => d && typeof d.id === 'number'},
-			{name: 'GET /api/top20', path: '/api/top20', check: (d: any) => Array.isArray(d)},
+			{
+				name: 'GET /api/runs',
+				path: '/api/runs',
+				check: (d: unknown) =>
+					isObject(d) && typeof d.total === 'number' && Array.isArray(d.runs),
+			},
+			{
+				name: `GET /api/runs/${testId}`,
+				path: `/api/runs/${testId}`,
+				check: (d: unknown) => isObject(d) && typeof d.id === 'number',
+			},
+			{
+				name: 'GET /api/top20',
+				path: '/api/top20',
+				check: (d: unknown) => Array.isArray(d),
+			},
 		]
 
 		for (const t of tests) {
 			try {
 				const res = await fetch(`${base}${t.path}`)
-				const data = await res.json()
+				const data: unknown = await res.json()
 				const ok = t.check(data)
 				console.log(`${ok ? '✓' : '✗'} ${t.name}`)
-				if (!ok) console.log('  Unexpected response:', JSON.stringify(data).slice(0, 100))
-			} catch (e) {
+				if (!ok)
+					console.log(
+						'  Unexpected response:',
+						JSON.stringify(data).slice(0, 100),
+					)
+			} catch (err) {
 				console.log(`✗ ${t.name}`)
-				console.log(`  Error: ${e}`)
+				console.log(`  Error: ${err}`)
 			}
 		}
 
 		// Test CORS preflight
 		try {
-			const res = await fetch(`${base}/api/runs`, {method: 'OPTIONS'})
+			const res = await fetch(`${base}/api/runs`, { method: 'OPTIONS' })
 			const cors = res.headers.get('access-control-allow-origin')
 			console.log(`${cors ? '✓' : '✗'} OPTIONS /api/runs (CORS)`)
-		} catch (e) {
+		} catch {
 			console.log(`✗ OPTIONS /api/runs (CORS)`)
 		}
 	},
