--- pages/_app.js
@@ -1,5 +1,0 @@
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
-}
-
-export default MyApp
