--- pages/index.js
@@ -1,14 +1,0 @@
-import Head from 'next/head'
-
-export default function Home() {
-  return (
-    <div>
-      <Head>
-        <title>Slay the Web API</title>
-        <link rel="icon" href="data:;base64,iVBORw0KGgo=" />
-      </Head>
-      <p>Are you lost? Most probably.</p>
-      <p>Fear not, visit <a href="https://github.com/oskarrough/slaytheweb-backend">github.com/oskarrough/slaytheweb-backend</a></p>
-    </div>
-  )
-}
