+++ wrangler.toml
@@ -1,6 +1,6 @@
 name = "slaytheweb-api"
-main = "worker/index.js"
-compatibility_date = "2024-01-01"
+main = "worker.js"
+compatibility_date = "2026-01-25"
 
 # TURSO_URL and TURSO_TOKEN should be set as secrets:
 # wrangler secret put TURSO_URL
